Upgrade greatexpections v1.0.0 and fix unit tests
