+++ pyproject.toml
@@ -23,7 +23,7 @@ include = [
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.12"
-great-expectations = "^0.18.20,<0.19" # <0.19 instead of <v1.0 to include v1 pre-releases
+great-expectations = ">=1.0.0"
 pydantic = ">=2.8.1,<3"
 pika = "^1.3.1"
 # needed for metrics serialization
